# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ServoMap API
  description: Real-time fuel prices across Australia. Aggregates all state government fuel APIs into one unified source.
  version: 1.0.0
  contact:
    name: ServoMap
    url: https://servo-map.com
servers:
  - url: https://api.servo-map.com/api/v1
    description: Production (api.servo-map.com)
  - url: http://localhost:8787/api/v1
    description: Local development
...
paths:
...
  /vehicles:
    get:
      operationId: searchVehicles
      tags: [Reference]
      summary: Search the car catalogue
      description: |
        The shared car catalogue used by "My car": model generations sold in Australia with body
        type, recommended fuel and fuel-tank capacity, each with the source it was read from.
        Every word of `q` must match the make or model; a four-digit word matches a model year.
      parameters:
        - name: q
          in: query
          description: Search words, e.g. `corolla 2020`. Empty returns the first `limit` models.
          schema:
            type: string
            maxLength: 60
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Matching model generations, by make, model and newest first
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Vehicle"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /me/fillups:
    put:
      operationId: putFillUps
      tags: [Accounts]
      summary: Add or replace fill-ups by id
      description: Idempotent; uploading the same fill-up twice leaves one copy. Invalid fields are all reported in `errors`.
      security:
        - session: []
      responses:
        "204":
          description: Stored
        "400":
          $ref: "#/components/responses/BadRequest"
  /me/fillups/{id}:
    delete:
      operationId: deleteFillUp
      tags: [Accounts]
      summary: Delete one fill-up
      security:
        - session: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: No such fill-up for this account
...
  /me/devices/{token}:
    put:
      operationId: putDevice
      tags: [Accounts]
      summary: Register this device's APNs token for alerts
      security:
        - session: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            pattern: "^[0-9a-f]{64,200}$"
      responses:
        "204":
          description: Registered
        "400":
          $ref: "#/components/responses/BadRequest"
components:
...
  securitySchemes:
    session:
      type: http
      scheme: bearer
      description: The session token from `POST /auth/{provider}`.
  schemas:
...
    Vehicle:
      type: object
      required: [id, make, model, fromYear, body, fuel, tankLitres, source, image]
      properties:
        id:
          type: string
          example: toyota-corolla-2019
        make:
          type: string
          example: Toyota
        model:
          type: string
          example: Corolla
        fromYear:
          type: integer
          example: 2019
        toYear:
          type: integer
          description: Last model year; absent while still on sale.
        body:
          type: string
          enum: [hatch, sedan, wagon, suv, ute, van]
        fuel:
          $ref: "#/components/schemas/FuelType"
        tankLitres:
          type: integer
          example: 50
        source:
          type: string
          format: uri
          description: Where the tank capacity was read.
        image:
          type: string
          example: /cars/generic-suv.jpg
          description: |
            Site-relative path of the car's picture on the website (resolve it against
            https://www.servo-map.com): the generation's own render when it has one, otherwise the
            unbranded stand-in for its body type, `/cars/generic-{body}.jpg`. An illustration, not a
            photo of the actual vehicle. Owned by `vehicleImagePath` in `@servo-map/shared`.
    FuelType:
      type: string
      enum: [U91, U95, U98, E10, Diesel]
      description: Fuel type code
...
    SuccessEnvelope:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [success]
    ErrorResponse:
      type: object
      required: [status, message, code]
      properties:
        status:
          type: string
          enum: [error]
        message:
          type: string
          description: Human-readable error description
        code:
          type: string
          description: Machine-readable error code
  responses:
    BadRequest:
      description: |
        Invalid request parameters. `code` is one of:
        `INVALID_STATE` (a `state` code is not a valid `AustralianState`),
        `INVALID_FUEL` (`fuel` is not a valid `FuelType`, or is missing where required), or
        `INVALID_GEO` (`lat`/`lng`/`radius` were not all provided together).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status: error
            message: "lat, lng, and radius must all be provided"
            code: INVALID_GEO
...